+++ backend/src/domain/repositories/recipe/in_memory/mod.rs
@@ -1,6 +1,5 @@
 use async_trait::async_trait;
 use eyre::eyre;
-use once_cell::sync::Lazy;
 use std::{
     collections::HashMap,
     sync::{Arc, Mutex},
